fix compilation on 32-bit targets
The magic number for bpffs doesn't fit in Statfs_t.Type on 32bit arch,
which leads to a compilation error. Looking at the kernel source, the
thing to do is to treat the magic number as unsigned.

Fixes #18

+++ syscalls.go
@@ -282,7 +282,7 @@ func bpfPinObject(fileName string, fd *bpfFD) error {
 	if err := unix.Statfs(dirName, &statfs); err != nil {
 		return err
 	}
-	if statfs.Type != bpfFSType {
+	if uint64(statfs.Type) != bpfFSType {
 		return errors.Errorf("%s is not on a bpf filesystem", fileName)
 	}
